requirements.txt
PyJWT==1.4.2
pycrypto==2.6.1
pem==19.2.0
requests==2.22.0